+++ src/reduce/reduce_multi.c
@@ -5,10 +5,12 @@
 #define N 1024
 #define NTHREADS 4
 
-void printMatrix(double **mat, int len)
+void printMatrix(float **mat, int len)
 {
-    for (int i = 0; i < len; i++) {
-        for (int j = 0; j < len; j++) {
+    for (int i = 0; i < len; i++)
+    {
+        for (int j = 0; j < len; j++)
+        {
             printf("\t%f", mat[i][j]);
         }
         printf("\n");
@@ -16,11 +18,11 @@ void printMatrix(double **mat, int len)
 }
 int main(int argc, char **argv)
 {
-srand(1);
-    
+    srand(1);
+
     int i, j, k, n;
     int nThreads;
-    double somma = 0;
+    float somma = 0;
     if (argc > 1)
     {
         n = atoi(argv[1]);
@@ -32,15 +34,15 @@ srand(1);
         nThreads = NTHREADS;
     }
 
-    double **a, **b, **c;
+    float **a, **b, **c;
 
-    a = (double **)malloc(n * sizeof(double *));
-    b = (double **)malloc(n * sizeof(double *));
+    a = (float **)malloc(n * sizeof(float *));
+    b = (float **)malloc(n * sizeof(float *));
 
     for (int x = 0; x < n; x++)
     {
-        a[x] = malloc(n * sizeof(double));
-        b[x] = malloc(n * sizeof(double));
+        a[x] = malloc(n * sizeof(float));
+        b[x] = malloc(n * sizeof(float));
     }
 
     //initialization
@@ -48,17 +50,16 @@ srand(1);
     {
         for (j = 0; j < n; j++)
         {
-            a[i][j] = ((double) rand()*(5)/(double)RAND_MAX-2);
-            b[i][j] = ((double) rand()*(5)/(double)RAND_MAX-2);
-            
+            a[i][j] = ((float)rand() * (5) / (float)RAND_MAX - 2);
+            b[i][j] = ((float)rand() * (5) / (float)RAND_MAX - 2);
         }
     }
-    //printf("RANDOM: %f\n",(double)rand()/ 5 - 2.0);
+    //printf("RANDOM: %f\n",(float)rand()/ 5 - 2.0);
 
     //calculate prod
     double begin = omp_get_wtime();
-    #pragma omp parallel for num_threads(nThreads) private(i,j) reduction(+ \
-                                                                        : somma)
+#pragma omp parallel for num_threads(nThreads) private(i, j) reduction(+ \
+                                                                       : somma)
     for (i = 0; i < n; i++)
     {
         for (j = 0; j < n; j++)
@@ -74,5 +75,5 @@ srand(1);
     printMatrix(b,n);
     */
     printf("Result sum: %f\n", somma);
-    printf("Time exec: %f sec, Matrix size: %d, Number Threads: %d\n", time_spent,n,nThreads);
+    printf("Time exec: %f sec, Matrix size: %d, Number Threads: %d\n", time_spent, n, nThreads);
 }
